Split hwIo into a package, moving the Hid classes into their own hwIo.hid module. Bulk and Serial could be moved into their own modules at a later date.
